requirements.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.34.0
accelerate>=0.24.0
pillow>=10.0.0
flask>=2.0.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
